+++ packages/sbtc/package.json
@@ -1,6 +1,6 @@
 {
   "name": "sbtc",
-  "version": "0.1.7",
+  "version": "0.1.8",
   "description": "Library for sBTC.",
   "license": "MIT",
   "author": "Hiro Systems PBC (https://hiro.so)",
@@ -27,7 +27,6 @@
     "@stacks/common": "^6.7.0",
     "@stacks/encryption": "^6.7.0",
     "@stacks/transactions": "^6.7.0",
-    "@stacks/wallet-sdk": "^6.7.0",
     "c32check": "^2.0.0",
     "micro-packed": "^0.3.2"
   },
